modification de la methode soustration pour qu elle puisse prendre en compte les données en milliard

+++ src/com/objis/cameroun/service/Service.java
@@ -21,9 +21,9 @@ public class Service implements IService {
 
 	/* Cette methode devrait etre implementée par Frank, mais par soucis de connexion internet, Mr Fouomene a implémenté*/
 	@Override
-	public int soustration(int a, int b) {
+	public double soustration(double a, double b) {
 		
-		int resultat;
+		double resultat;
 		
 		resultat = a - b;
